Python Image Retrieval Server/requirements.txt
Pillow
h5py
tensorflow
Keras
Flask